requirements.txt
dj-database-url==2.1.0
django-ckeditor==6.7.1
psycopg2-binary==2.9.10
asgiref>=3.9.1
cloudinary==1.44.1
Django==6.0.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
